pyproject.toml
[project]
name = "gatherarr"
version = "0.1.0"
description = "A minimal worker-style Arr search runner (initial skeleton)."
requires-python = "==3.14.*"
license = { text = "MIT" }
dependencies = [
  "attrs>=25.4.0",
  "flask>=3.0.0",
  "httpx>=0.25.0",
  "prometheus-client>=0.19.0",
  "pydantic-settings>=2.0.0",
  "python-dateutil>=2.8.0",
  "pyyaml>=6.0.1",
  "structlog>=24.1.0",
  "tenacity>=8.2.0",
]

[dependency-groups]
dev = [
  "mypy>=1.7.0",
  "openapi-core>=0.22.0",
  "openapi-python-client>=0.28.2",
  "poethepoet>=0.24.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "pytest-timeout>=2.2.0",
  "ruff>=0.1.0",
  "types-python-dateutil>=2.9.0",
  "types-pyyaml>=6.0.12",
  "types-requests>=2.32.4.20260107",
]

[tool.poe.tasks]
check = ["format", "lint", "type-check", "test"]
check-e2e = ["check", "test-e2e"]
generate-clients = "openapi-python-client generate --path context/radarr_api.json --output-path app/radarr_client --meta none --overwrite && openapi-python-client generate --path context/sonarr_api.json --output-path app/sonarr_client --meta none --overwrite"
format = "ruff format ."
format-check = "ruff format --check ."
lint = "ruff check --fix ."
lint-check = "ruff check ."
type-check = "mypy ."
test = "pytest tests --ignore=tests/integration"
test-e2e = "pytest tests/integration"

[tool.ruff]
target-version = "py314"
line-length = 100
indent-width = 2

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
exclude = ["app/radarr_client", "app/sonarr_client"]
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
follow_imports = "normal"
ignore_missing_imports = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
  "-v",
]
timeout = 60
timeout_method = "thread"
